Divide ConsultantRow into new files for each component
