+++ frontend/src/components/Staffing/ConsultantRow.tsx
@@ -1,0 +1,94 @@
+"use client";
+import { Consultant } from "@/types";
+import React, { useState } from "react";
+import { ChevronDown } from "react-feather";
+import { DetailedBookingRows } from "@/components/Staffing/DetailedBookingRows";
+import { WeekCell } from "@/components/Staffing/WeekCell";
+import { AddStaffingCell } from "@/components/Staffing/AddStaffingCell";
+
+export default function ConsultantRows({
+  consultant,
+}: {
+  consultant: Consultant;
+}) {
+  const [isListElementVisible, setIsListElementVisible] = useState(false);
+  const [isRowHovered, setIsRowHovered] = useState(false);
+  const [hoveredRowWeek, setHoveredRowWeek] = useState(-1);
+
+  const columnCount = consultant.bookings.length ?? 0;
+
+  function toggleListElementVisibility() {
+    setIsListElementVisible(!isListElementVisible);
+  }
+
+  return (
+    <>
+      <tr
+        className="h-[52px]"
+        onMouseEnter={() => setIsRowHovered(true)}
+        onMouseLeave={() => setIsRowHovered(false)}
+      >
+        <td
+          className={`border-l-2 ${
+            isListElementVisible
+              ? "border-l-secondary"
+              : isRowHovered
+              ? "border-l-primary"
+              : "border-l-primary/5"
+          } `}
+        >
+          <button
+            className={`p-2 rounded-lg ml-2 hover:bg-primary hover:bg-opacity-10 ${
+              isListElementVisible && "rotate-180"
+            }`}
+            onClick={toggleListElementVisibility}
+          >
+            <ChevronDown className={`text-primary w-6 h-6`} />
+          </button>
+        </td>
+        <td>
+          <div className="flex flex-col gap-1 ">
+            <p
+              className={`text-black text-start ${
+                isListElementVisible ? "normal-medium" : "normal"
+              }`}
+            >
+              {consultant.name}
+            </p>
+            <p className="xsmall text-black/75 text-start">
+              {`${consultant.yearsOfExperience} års erfaring`}
+            </p>
+          </div>
+        </td>
+        {consultant.bookings?.map((b, index) => (
+          <WeekCell
+            key={index}
+            bookedHoursPerWeek={b}
+            isListElementVisible={isListElementVisible}
+            setIsListElementVisible={setIsListElementVisible}
+            consultant={consultant}
+            setHoveredRowWeek={setHoveredRowWeek}
+            hoveredRowWeek={hoveredRowWeek}
+            columnCount={columnCount}
+            isLastCol={index == consultant.bookings.length - 1}
+            isSecondLastCol={index == consultant.bookings.length - 2}
+          />
+        ))}
+      </tr>
+      {isListElementVisible &&
+        consultant.detailedBooking &&
+        consultant.detailedBooking.map((db, index) => (
+          <DetailedBookingRows
+            key={index}
+            consultant={consultant}
+            detailedBooking={db}
+          />
+        ))}
+      {isListElementVisible && (
+        <tr>
+          <AddStaffingCell />
+        </tr>
+      )}
+    </>
+  );
+}
