+++ frontend/src/components/Staffing/InfoPill.tsx
