+++ frontend/src/pagecontent/StaffingContent.tsx
@@ -2,7 +2,7 @@
 
 import StaffingSidebar from "@/components/StaffingSidebar";
 import FilteredConsultantsList from "@/components/FilteredConsultantsList";
-import InfoPillDescriptions from "@/components/InfoPillDescriptions";
+import InfoPillDescriptions from "@/components/Staffing/InfoPillDescriptions";
 import { useState } from "react";
 import IconActionButton from "@/components/Buttons/IconActionButton";
 import { Filter } from "react-feather";
